+++ storage/src/main/java/tech/pegasys/teku/storage/store/StoreConfig.java
@@ -26,20 +26,25 @@ public class StoreConfig {
   public static final int DEFAULT_CHECKPOINT_STATE_CACHE_SIZE = 20;
   public static final int DEFAULT_HOT_STATE_PERSISTENCE_FREQUENCY_IN_EPOCHS = 2;
 
+  public static final int DEFAULT_EARLIEST_AVAILABLE_BLOCK_SLOT_QUERY_FREQUENCY = 0;
+
   private final int stateCacheSize;
   private final int blockCacheSize;
   private final int checkpointStateCacheSize;
   private final int hotStatePersistenceFrequencyInEpochs;
+  private final int earliestAvailableBlockSlotFrequency;
 
   private StoreConfig(
       final int stateCacheSize,
       final int blockCacheSize,
       final int checkpointStateCacheSize,
-      final int hotStatePersistenceFrequencyInEpochs) {
+      final int hotStatePersistenceFrequencyInEpochs,
+      int earliestAvailableBlockSlotFrequency) {
     this.stateCacheSize = stateCacheSize;
     this.blockCacheSize = blockCacheSize;
     this.checkpointStateCacheSize = checkpointStateCacheSize;
     this.hotStatePersistenceFrequencyInEpochs = hotStatePersistenceFrequencyInEpochs;
+    this.earliestAvailableBlockSlotFrequency = earliestAvailableBlockSlotFrequency;
   }
 
   public static Builder builder() {
@@ -60,6 +65,10 @@ public class StoreConfig {
 
   public int getCheckpointStateCacheSize() {
     return checkpointStateCacheSize;
+  }
+
+  public int getEarliestAvailableBlockSlotFrequency() {
+    return earliestAvailableBlockSlotFrequency;
   }
 
   public int getHotStatePersistenceFrequencyInEpochs() {
@@ -96,6 +105,7 @@ public class StoreConfig {
     private int checkpointStateCacheSize = DEFAULT_CHECKPOINT_STATE_CACHE_SIZE;
     private int hotStatePersistenceFrequencyInEpochs =
         DEFAULT_HOT_STATE_PERSISTENCE_FREQUENCY_IN_EPOCHS;
+    private int earliestAvailableBlockSlotFrequency = 0;
 
     private Builder() {}
 
@@ -104,7 +114,8 @@ public class StoreConfig {
           stateCacheSize,
           blockCacheSize,
           checkpointStateCacheSize,
-          hotStatePersistenceFrequencyInEpochs);
+          hotStatePersistenceFrequencyInEpochs,
+          earliestAvailableBlockSlotFrequency);
     }
 
     public Builder stateCacheSize(final int stateCacheSize) {
@@ -122,6 +133,12 @@ public class StoreConfig {
     public Builder checkpointStateCacheSize(final int checkpointStateCacheSize) {
       validateCacheSize(checkpointStateCacheSize);
       this.checkpointStateCacheSize = checkpointStateCacheSize;
+      return this;
+    }
+
+    public Builder earliestAvailableBlockSlotFrequency(
+        int earliestAvailableBlockSlotQueryFrequency) {
+      this.earliestAvailableBlockSlotFrequency = earliestAvailableBlockSlotQueryFrequency;
       return this;
     }
